Coerce "hub pull-request" to properly detect the tracking branch (#7)
When opening a pull request with "hub", the command will attempt to be
helpful and try to detect when the local branch contains commits that
are not present in the remote branch that'll be used for the pull
request. This is so that it's more difficult to open a pull request,
only to find out it doesn't contain all fo the intended commits; e.g.
because some commits weren't pushed to the remote branch.

In order for this check to work properly in our case, we need to ensure
it's checking the branch we push to under the "sync-with-upstream"
namespace. We (now) do this by setting the local branch's tracking
branch to point to that remote branch, and also by setting the
"push.default" git configuration tunable. The "hub" command currently
checks that tunable and modifies it's behavior based on that value.

By default, the "hub" command would detect a remote branch with the same
name as the local branch, and assume that branch was the local branch's
remote tracking branch; which is not the case for this script, and could
result in "hub pull-request" throwing an error like the following:

    + hub pull-request -F - -b master -h projects/sync-with-upstream/master
    Aborted: 1 commits are not yet pushed to origin/master
    (use `-f` to force submit a pull request anyway)