sync-with-upstream: do nothing if workflow PR already exists
We recently noticed that the specific workflow fails once a PR
for it has already been opened but hasn't been merged. The
problem is that with the current branch protections standards
of Delphix repos on Github we are not allowed to force-push to
the downstream branch. Assuming no re-writes in the history of
the upstream repo, this commit avoids force-pushing in this
workflow by only pushing code to the downstream branch when
there is no open PR for it.