Use "gh" and "auto-merge" feature of PRs
This change replaces the use of the "hub" command with "gh", since gh is
now the "official" CLI tool supported by GitHub. Additionally, after
opening the PR, we now enable auto-merge on the it; this way, depending
on how the PR checks and review requirements are configured, the PR can
potentially land automatically without human intervention.